backend/scripts/repo_B/requirements.txt
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.0
PyYAML==5.3.1
scikit-learn==0.22.2
scipy==1.5.0
xlrd==1.2.0
